requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.0
supabase==2.10.0
websockets>=13,<16
